No-running of map V3.1
Fox & Hedgehog are not running out of the map anymore. And telepoting to X & Y pos 150 in map if running out of the map.

Experimented in TEST MODE, IN PROGRESS.py

Notes:

If you want to play V3.1 Instead of new versions, Just download V3.1.py
TEST MODE, IN PROGRESS.py expands on through V3.1 and expanding it's amazing progress into an awesome game.